+++ src/Handlers/ErrorHandler.php
@@ -1,0 +1,40 @@
+<?php
+/**
+ * COPS (Calibre OPDS PHP Server) class file
+ *
+ * @license    GPL v2 or later (http://www.gnu.org/licenses/gpl.html)
+ * @author     Sébastien Lucas <sebastien@slucas.fr>
+ * @author     mikespub
+ */
+
+namespace SebLucas\Cops\Handlers;
+
+use SebLucas\Cops\Output\Response;
+
+/**
+ * Summary of ErrorHandler
+ */
+class ErrorHandler extends BaseHandler
+{
+    public const HANDLER = "error";
+
+    public static function getRoutes()
+    {
+        return [];
+    }
+
+    public function handle($request)
+    {
+        if ($request->getHandler() == 'index' && $request->path() != '' && !$request->isJson()) {
+            //Response::redirect(Route::link("index", null, ["db" => $request->database()]));
+            $error = "Invalid request path '" . $request->path() . "'";
+            $ref = $request->server('HTTP_REFERER');
+            if ($ref) {
+                $error .= ' from ' . $ref;
+            }
+            Response::sendError($request, $error, ["db" => $request->database()]);
+            return;
+        }
+        Response::notFound($request);
+    }
+}